@@ -1,33 +1,46 @@
 {
-  "authors": [
-    {
-      "name": "Eric Richer",
-      "email": "eric.richer@vistoconsulting.com",
-      "role": "Lead"
+    "name": "visto9259/laminas-webpack",
+    "authors": [
+        {
+            "name": "Eric Richer",
+            "email": "eric.richer@vistoconsulting.com",
+            "role": "Lead"
+        }
+    ],
+    "autoload": {
+        "psr-4": {
+            "Webpack\\": "src"
+        }
+    },
+    "description": "Laminas Webpack integration",
+    "keywords": [
+        "laminas",
+        "webpack"
+    ],
+    "license": "MIT",
+    "require": {
+        "php": "^7.4 || ~8.0.0 || ~8.1.0",
+        "laminas/laminas-stdlib": "^3.2",
+        "laminas/laminas-modulemanager": "^2.8",
+        "laminas/laminas-mvc": "^3.1",
+        "laminas/laminas-view": "^2.11",
+        "laminas/laminas-eventmanager": "^3.2",
+        "laminas/laminas-servicemanager": "^3.4"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^9.5",
+        "phpspec/prophecy-phpunit": "^2.0",
+        "mikey179/vfsstream": "^1.6.11",
+        "laminas/laminas-coding-standard": "^2.3"
+    },
+    "config": {
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
+    },
+    "scripts": {
+        "test": ["@phpcs", "@phpunit"],
+        "phpunit": "phpunit --colors",
+        "phpcs": "phpcs --colors"
     }
-  ],
-  "autoload": {
-    "psr-4": {
-      "Webpack\\": "src"
-    }
-  },
-  "description": "Laminas Webpack integration",
-  "keywords": [
-    "laminas",
-    "webpack"
-  ],
-  "license": "MIT",
-  "name": "visto9259/laminas-webpack",
-  "require": {
-    "php": "^7.3 || ^8.0",
-    "laminas/laminas-stdlib": "^3.2",
-    "laminas/laminas-modulemanager": "^2.8",
-    "laminas/laminas-mvc": "^3.1",
-    "laminas/laminas-view": "^2.11",
-    "laminas/laminas-eventmanager": "^3.2",
-    "laminas/laminas-servicemanager": "^3.4"
-  },
-  "require-dev": {
-  },
-  "version": "1.1.0"
 }
