@@ -1,23 +1,23 @@
 <?php
 
+declare(strict_types=1);
 
 namespace Webpack\Listener;
 
-
-use Exception;
-use Interop\Container\ContainerInterface;
-use Webpack\Config\WebpackOptions;
 use Laminas\ServiceManager\Factory\FactoryInterface;
+use Psr\Container\ContainerInterface;
+use Webpack\Config\WebpackOptions;
 
 class WebpackRouteListenerFactory implements FactoryInterface
 {
-
     /**
      * @inheritDoc
-     * @throws Exception
      */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
-    {
+    public function __invoke(
+        ContainerInterface $container,
+        $requestedName,
+        ?array $options = null
+    ): WebpackRouteListener {
         $webpackOptions = $container->get(WebpackOptions::class);
         return new WebpackRouteListener($webpackOptions);
     }
