@@ -1,22 +1,19 @@
 <?php
 
+declare(strict_types=1);
 
 namespace Webpack\Listener;
 
-
-use Interop\Container\ContainerInterface;
-use Interop\Container\Exception\ContainerException;
-use Laminas\ServiceManager\Exception\ServiceNotCreatedException;
-use Laminas\ServiceManager\Exception\ServiceNotFoundException;
+use Laminas\ServiceManager\Factory\FactoryInterface;
+use Psr\Container\ContainerInterface;
 use Webpack\Config\WebpackOptions;
 
-class WebpackViewListenerFactory implements \Laminas\ServiceManager\Factory\FactoryInterface
+class WebpackViewListenerFactory implements FactoryInterface
 {
-
     /**
      * @inheritDoc
      */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    public function __invoke(ContainerInterface $container, $requestedName, ?array $options = null): WebpackViewListener
     {
         return new WebpackViewListener($container->get(WebpackOptions::class));
     }
